Refactor so I can possibly close the db connection easily
